move js script to webapp folder
This allows the browser to load the map files thus making it easier to
step through the code for debugging
