+++ .gitignore
@@ -9,3 +9,4 @@ test-output/
 *.iml
 node/
 node_modules/
+src/main/webapp/js
