+++ src/main/resources/org/biouno/unochoice/DynamicReferenceParameter/index.jelly
@@ -2,7 +2,7 @@
 <j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:f="/lib/form" xmlns:l="/lib/layout">
   ${it.parameters.clear()}
   <st:adjunct includes="io.jenkins.plugins.jquery3"/>
-  <st:adjunct includes="org.biouno.unochoice.stapler.unochoice"/>
+  <st:include page="/org/biouno/unochoice/stapler/unochoice.jelly"/>
   <j:set var="choiceType" value="${it.choiceType}"/>
   <j:set var="omitValueField" value="${it.omitValueField}"/>
   <j:set var="cssclazz" value="" />
