+++ src/main/webapp/css/unochoice.css
@@ -1,0 +1,17 @@
+.js-loading-overlay {
+  position:absolute;
+  top:0; bottom:0;
+  left:0; right:0;
+
+  z-index:199;
+
+  background:#fff;
+  opacity:0.5;
+
+  border-radius: inherit;
+}
+.active-choice {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+}
